fastapi_rbas/app/requirements.txt
fastapi
sqlalchemy
pydantic
passlib
python-jose
alembic
uvicorn
python-dotenv
